requirements.txt
pillow==11.0.0
pyexiv2==2.15.0
